sink(ticdc): improve DDL timestamp failpoint coverage
