+++ tests/integration_tests/ddl_default_current_timestamp/run.sh
@@ -16,9 +16,11 @@ function run() {
 	cd $WORK_DIR
 
 	start_ts=$(run_cdc_cli_tso_query ${UP_PD_HOST_1} ${UP_PD_PORT_1})
+	# Force a single session and skip set/reset for targeted DDLs to simulate leakage.
 	export GO_FAILPOINTS='github.com/pingcap/tiflow/cdc/sink/ddlsink/mysql/MySQLSinkForceSingleConnection=return(true);github.com/pingcap/tiflow/cdc/sink/ddlsink/mysql/MySQLSinkSkipResetSessionTimestampAfterDDL=return("c2");github.com/pingcap/tiflow/cdc/sink/ddlsink/mysql/MySQLSinkSkipSetSessionTimestamp=return("d2")'
 	run_cdc_server --workdir $WORK_DIR --binary $CDC_BINARY
 	run_cdc_cli changefeed create --start-ts=$start_ts --sink-uri="mysql://normal:123456@127.0.0.1:3306/"
+	export GO_FAILPOINTS=''
 
 	run_sql "DROP DATABASE IF EXISTS ${DB_NAME};" ${UP_TIDB_HOST} ${UP_TIDB_PORT}
 	run_sql "CREATE DATABASE ${DB_NAME};" ${UP_TIDB_HOST} ${UP_TIDB_PORT}
@@ -76,6 +78,8 @@ function run() {
 if [ "$SINK_TYPE" == "mysql" ]; then
 	trap stop_tidb_cluster EXIT
 	run $*
+	check_logs_contains $WORK_DIR "Skip setting session timestamp due to failpoint"
+	check_logs_contains $WORK_DIR "Skip resetting session timestamp after DDL execution due to failpoint"
 	check_logs $WORK_DIR
 	echo "[$(date)] <<<<<< run test case $TEST_NAME success! >>>>>>"
 fi
